Moved timings to App because multiple children need current turn. Make past turns grey.
