+++ src/components/PlanetEdit.vue
@@ -224,7 +224,6 @@
         </td>
       </tr>
     </table>
-
   </border-box>
 </template>
