+++ src/scss/template/_shared.scss
@@ -171,3 +171,7 @@ ul.draggable li {
     color: #FFFFFF;
     text-align: center;
 }
+
+.grey {
+    color: grey;
+}
